+++ packages/server/src/db/models/Word.ts
@@ -7,13 +7,17 @@ import {
   Word as WordType,
   NewWordInput,
   UpdateStatisticsInput,
-  Language,
   Game as GameType,
-  SortBy,
-  SortWordsBy,
-  WordStatisticsField
+  WordStatisticsField,
+  PaginatedWords
 } from '../../generated/graphql';
-import { formatFilter, formatData } from '../helpers';
+import {
+  formatFilter,
+  formatData,
+  getWordsFilters,
+  WordsFilterType,
+  WordsWithPaginationFilter
+} from '../helpers';
 
 const STATISTICS_FIELD: WordStatisticsField = {
   lastTimePracticed: 0,
@@ -26,17 +30,8 @@ const DEFAULT_STATISTICS: Record<GameType, WordStatisticsField> = {
   [GameType.Audio]: STATISTICS_FIELD,
   [GameType.SelectDef]: STATISTICS_FIELD,
   [GameType.SelectWord]: STATISTICS_FIELD,
-  [GameType.TypeWord]: STATISTICS_FIELD
-};
-
-type wordsFilter = {
-  sortBy?: SortBy | SortWordsBy | 'updatedAt';
-  language: Language;
-  isReverseOrder: boolean;
-  timesToLearn?: number | null;
-  gameType?: GameType;
-  user?: string;
-  tags?: string[];
+  [GameType.TypeWord]: STATISTICS_FIELD,
+  [GameType.Conjugation]: STATISTICS_FIELD
 };
 
 export interface WordModelType {
@@ -46,7 +41,10 @@ export interface WordModelType {
     projection: string,
     options?: QueryOptions
   ) => Promise<WordType[] | null>;
-  findManyAndSort: (filter: wordsFilter) => Promise<WordType[] | null>;
+  findManyAndSort: (filter: WordsFilterType) => Promise<WordType[] | null>;
+  findManyAndPaginate: (
+    filter: WordsWithPaginationFilter
+  ) => Promise<PaginatedWords | null>;
   createOne: (fields: NewWordInput) => Promise<WordType | null>;
   updateOne: (
     fields: Partial<WordType> & Pick<WordType, 'id' | 'user'>
@@ -86,83 +84,14 @@ export const WordModel: WordModelType = {
     return words;
   },
 
-  // Check if criteria is defined, if so => sort by criteria, else use natural sorting
   async findManyAndSort(filter) {
-    const {
-      user,
-      language = Language.English,
-      sortBy,
-      isReverseOrder,
-      gameType,
-      tags,
-      timesToLearn = 5
-    } = filter;
+    const { user } = filter;
 
     if (!user) {
       return [];
     }
 
-    let orderNum = isReverseOrder ? -1 : 1;
-
-    // words with the highest number of errors should go first
-    if (sortBy === SortBy.ErrorCount) {
-      orderNum = -1 * orderNum;
-    }
-    let learningStatusFilter = {};
-    let sort: Record<string, number> = { $natural: orderNum };
-
-    let tagsFilters: { tags: string }[] = [];
-    if (tags?.length) {
-      tagsFilters = tags?.map(tag => {
-        return { tags: tag };
-      });
-    }
-    // TODO: refactor this part
-    // Select words that are not learned or have been practiced without error less than 5 times in a row
-    if (gameType && sortBy !== SortBy.MemoryRefresher) {
-      learningStatusFilter = {
-        $and: [
-          { isLearned: { $ne: true } },
-          ...tagsFilters,
-          {
-            $or: [
-              { [`statistics.${gameType}.successRate`]: { $lt: timesToLearn } },
-              { [`statistics.${gameType}.successRate`]: null }
-            ]
-          }
-        ]
-      };
-    } else {
-      if (tagsFilters?.length) {
-        learningStatusFilter = {
-          $and: tagsFilters
-        };
-      }
-    }
-
-    if (sortBy) {
-      let propName: string = sortBy;
-      if (gameType) {
-        propName = `statistics.${gameType}.${sortBy}`;
-      }
-
-      if (sortBy === SortBy.MemoryRefresher) {
-        learningStatusFilter = {
-          $or: [
-            { [`statistics.${gameType}.successRate`]: { $gte: timesToLearn } },
-            { isLearned: true }
-          ]
-        };
-      } else {
-        sort = { [propName]: orderNum };
-      }
-    }
-
-    const filters = {
-      user,
-      language,
-      ...learningStatusFilter
-    };
+    const { sort, filters } = getWordsFilters(filter);
 
     const words = await Word.find(filters)
       // @ts-ignore
@@ -171,6 +100,38 @@ export const WordModel: WordModelType = {
       .exec();
 
     return words;
+  },
+
+  async findManyAndPaginate(filter) {
+    const { page = 1, limit = 5 } = filter;
+    const docsLimit = Math.abs(limit);
+    const skip = (Math.abs(page) - 1) * docsLimit;
+    const { user } = filter;
+
+    if (!user) {
+      return null;
+    }
+
+    const { sort, filters } = getWordsFilters(filter);
+    const wordsCount = await Word.countDocuments(filters);
+    const hasNext = wordsCount - Math.abs(page) * limit > 0;
+    const words = await Word.find(filters)
+      // @ts-ignore
+      .sort(sort)
+      .limit(docsLimit)
+      .skip(skip)
+      .populate('tags')
+      .exec();
+
+    const result: PaginatedWords = {
+      words,
+      page,
+      limit,
+      hasNext,
+      wordsCount
+    };
+
+    return result;
   },
 
   async createOne(fields) {
@@ -245,6 +206,10 @@ export const WordModel: WordModelType = {
         const timesToLearn = gameConfig?.timesToLearn || 5;
 
         const currentStatistics = word?.statistics?.[gameType];
+
+        if (!currentStatistics) {
+          word.statistics[gameType] = DEFAULT_STATISTICS[gameType];
+        }
 
         const newStatistics = {
           practicedTimes: 1,