+++ packages/server/src/resolvers/Query.ts
@@ -23,7 +23,7 @@ export const QueryResolvers: QueryResolversType<ResolverContext> = {
       languages: lang
     });
     if (!games) {
-      throw new UserInputError(`no games for ${language} language were found`);
+      throw new UserInputError(`no games for ${lang} language were found`);
     }
     return games;
   },
@@ -60,7 +60,8 @@ export const QueryResolvers: QueryResolversType<ResolverContext> = {
         gameType,
         sortBy,
         isReverseOrder,
-        tags
+        tags,
+        tense = 'pind'
       } = input;
       const config = await models.Game.findOne({
         type: gameType,
@@ -107,14 +108,28 @@ export const QueryResolvers: QueryResolversType<ResolverContext> = {
         );
       }
 
-      return generateGameData(gameType, words, config, optionsMaterial);
+      return generateGameData(gameType, words, config, optionsMaterial, tense);
     }
   ),
-  tags: authenticated(async (_, { language }, { models, user }) => {
-    const tags = await models.WordTag.findMany({
-      language,
-      user: user?.id
+  tags: authenticated(
+    async (_, { language = Language.English }, { models, user }) => {
+      const tags = await models.WordTag.findMany({
+        language,
+        user: user?.id
+      });
+      return tags;
+    }
+  ),
+  wordsPerPage: authenticated(async (_, { input }, { models, user }) => {
+    const sortBy = input?.sortBy || 'updatedAt';
+    const isReverseOrder =
+      !input?.isReverseOrder && !input?.sortBy ? true : input?.isReverseOrder;
+    const result = await models.Word.findManyAndPaginate({
+      ...input,
+      user: user?.id,
+      sortBy,
+      isReverseOrder
     });
-    return tags;
+    return result;
   })
 };