+++ packages/ui/src/components/TagSelector/TagSelector.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 import { cn } from '../../utils/classnames';
 
 import styles from './TagSelector.module.css';
@@ -13,21 +13,38 @@ export interface TagSelectorProps {
   onChange: (val: string[]) => void;
   value?: string[];
   label?: string;
+  showNoTagsTag?: boolean;
   /**additional styling */
   className?: string;
 }
+
+// @ts-ignore
+const NO_TAGS_TAG: WordTag = {
+  color: '#F7F7F7',
+  text: 'without tags',
+  id: '000000000000000000000000',
+  desc: 'select words that have no tags'
+};
 /**Component description goes here */
 export const TagSelector = ({
-  tags,
+  tags: initialTags = [],
   value = [],
   className,
   onChange,
-  label
+  label,
+  showNoTagsTag
 }: TagSelectorProps) => {
   const handleChange = (val: string) => {
     const values = [...value, val];
     onChange(values);
   };
+
+  const tags = useMemo(() => {
+    if (showNoTagsTag) {
+      return [...initialTags, NO_TAGS_TAG];
+    }
+    return initialTags;
+  }, [initialTags, showNoTagsTag]);
 
   const { Select, Option } = useSelect<string>({
     onChange: handleChange,
@@ -70,7 +87,12 @@ export const TagSelector = ({
           }
           return (
             <li key={id}>
-              <Tag color={color} text={text} className={styles.tag}>
+              <Tag
+                color={color}
+                text={text}
+                className={styles.tag}
+                iconId={tag.id === NO_TAGS_TAG.id ? 'void' : undefined}
+              >
                 <Button
                   className={styles.button}
                   variant="icon"