+++ packages/front/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.1.1 (2024-07-02)
+
+### Bug Fixes
+
+- fix minor bugs ([#177](https://github.com/marusyaganza/lp-monorepo/issues/177)) ([db06206](https://github.com/marusyaganza/lp-monorepo/commit/db0620660b9672204613def0c00c6fef3b1d6817))
+- fix word form field width on desktop ([#178](https://github.com/marusyaganza/lp-monorepo/issues/178)) ([8d4a59f](https://github.com/marusyaganza/lp-monorepo/commit/8d4a59f6db2f1a9a583b3670d433a157d7a5f2d8))
+
 # 1.1.0 (2024-06-08)
 
 ### Features